Fixed synchronization problems by adjusting coordinates manually after the animation.
It's worth noting that the animation process left an animated object in a place with
slightly different coordinates from the ones being actually passed to animate() method.
I suppose this should be fixed in Raphael.

Also removed -0.005 constant in collision resolving (used instead of 0), which meant to be a
threshold for tolerating minor errors during calculations. It seems this lead to
subtle bugs on later stages.